Ignore ECHILD error on wait, avoid trying to reap child twice
On some OSes, setting SIGCHLD to SIG_IGN means that the OS will reap
the child automatically, so trying to wait on it returns ECHILD, so
instead of complaining we ignore that error.
Always set c->pid to 0 after waiting, so that we avoid waiting on
the child twice, ref.  http://gnats.netbsd.org/50048